Advancement Screen
The first upload of the Advancement Screen component. No AI changes,
just the popup for social policies and tenets, as well as the policy
branch classes (done through SQL). Mostly put here to keep it backed up
until the DLL changes to the AI and can be finished.

Still usable in this state, but does not allow as much modability, due
to the AI's limitations with Ideologies.